Remove strict types for compatibility
